config: split the junos-host name resolution out of the deny projection
The #6705 change pushed pkg/config/junos_host_deny.go from 1497 to 1532
LOC, crossing the 1500 [WATCH] modularity floor, so the audit gate reds
the branch that grew it.

None of the three documented "When NOT to refactor" cases applies here:
the file is not awaiting replacement, and it is not dominated by one
cohesive function or schema. It has a real seam. Everything moved is the
NAME-RESOLUTION half — turning the address, application and port tokens a
policy carries into concrete per-family CIDR sets and L4 tuples — and none
of it reads or writes projection state. The dependency runs one way: the
projection calls into these helpers and never the reverse.

Mechanical move of a contiguous range, cut at doc-comment boundaries so no
function inherits its neighbour's comment. No call site changed, no symbol
renamed, no behaviour depends on which file a package-internal name lives
in. junos_host_deny.go 1532 -> 1193; junos_host_deny_resolve.go 360.

Validation: go build and gofmt clean on both files, pkg/refactoraudit
green, and go test -count=1 still green across pkg/config, pkg/daemon,
pkg/nftables and pkg/dataplane/userspace.
